requirements.txt
appdirs
numpy
pyyaml
joblib
scipy
prettytable
pandas
h5py
